Working on LKF NEES, its a mess right now.
The ground truth is wrong and I need to work on fixing it. Right now it comes from propogation of the linear model, which is incorrect. I need to also fix how noise is added to the ground truth. 

The basic framework for the Monte Carlo simulation loops and NEES calculations is here.